+++ helpers/task.py
@@ -1,9 +1,4 @@
-"""contains the helper functions for the tasks and sessions"""
-
-import glob
-import importlib
-import sys
-from os.path import basename, dirname, isfile, join
+"""contains the helper class and functions for the tasks"""
 
 
 def task_to_list(tasklist):
@@ -76,28 +71,6 @@ class TaskBase:
         self.log(self.__doc__)
         self.log("################")
 
-    # @staticmethod
-    # def int_input(varname, debug=False):
-    #     """Get an integer input from the user."""
-    #     while True:
-    #         try:
-    #             return int(input(f"Enter {varname}: "))
-    #         except ValueError as err:
-    #             if debug:
-    #                 raise ValueError from err
-    #             print("Invalid input. Please enter an integer.")
-
-    # @staticmethod
-    # def float_input(varname, debug=False):
-    #     """Get a float input from the user."""
-    #     while True:
-    #         try:
-    #             return float(input(f"Enter {varname}: "))
-    #         except ValueError as err:
-    #             if debug:
-    #                 raise ValueError from err
-    #             print("Invalid input. Please enter a float.")
-
     def run_tasks(self):
         """Run all tasks in the tasklist."""
         for task in self.tasklist:
@@ -124,28 +97,3 @@ class TaskBase:
         var_str = "; ".join(var_list)
         print(msg, var_str)
         return f"{msg} {var_str}"
-
-
-def run_session(file, subdir=""):
-    """Run all tasks in all files in directory of the file.
-    Optional argument dir can be used to specify a child directory.
-    Ignores main.py and files starting with _."""
-
-    files = sorted(glob.glob(join(dirname(file), subdir, "*.py")))
-    for _file in files:
-        if not isfile(_file):
-            continue
-        fname = basename(_file)
-        if _file.endswith("main.py") or fname.startswith("_"):
-            continue
-        subdir = subdir.replace("/", ".").replace("\\", ".")
-        modname = f"{subdir}.{fname[:-3]}" if subdir else fname[:-3]
-        importlib.import_module(modname)
-        mod = sys.modules[modname]
-        try:
-            mod.Task(mod.__name__).run_tasks()
-        except AttributeError as err:
-            print(err)
-        except ValueError as err:
-            print(err)
-        print("\n")
